+++ components/core/ui/theme/src/main/kotlin/com/flipperdevices/core/ui/theme/models/FlipperPallet.kt
@@ -121,6 +121,7 @@ data class FlipperPallet(
 
     val shareSheetBackground: Color,
     val shareSheetBackgroundAction: Color = Color(0xFFFF8200),
+    val shareSheetScrimColor: Color,
     val fapHubSwitchBackground: Color,
     val fapHubSelectedBackgroundColor: Color,
     val fapHubActiveColor: Color = text100,