+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [Feature] Add application catalog list with sorted button
 - [FIX] Uploading share errors
 - [FIX] Display by white color meta data on NFC card
+- [FIX] Scrim status bar and design changes on bottom sheet
 
 # 1.3.0